+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+
+## Unreleased
+### Added
+- `\flipbox\ember\controllers\LogViewerTrait` to assist w/ consuming and digesting log files
+
 ## 1.0.7 - 2018-07-17
 ### Changed
 - `\flipbox\ember\services\traits\queries\Accessor::findAllByCondition` will attempt to set query property values prior to creating a 'where' condition.